Add files via upload
For this assignment, we were asked to use the PCam dataset from the Kaggle competition for Histopathologic Cancer Detection to conduct analysis of the images and build CNN models for binary classification.  The data was split into training and validation sets in order to monitor performance as we attempted to distinguish images with a cancerous tumor from those images with no cancerous tumor.  This python notebook has been uploaded in .pdf format.